finished roadmap restructure
