--- src/components/Main/main.module.scss
@@ -1,6 +1,0 @@
-.main {
-    padding-top: 32px;
-    padding-bottom: 56px;
-    padding-left: 24px;
-    padding-right: 24px;
-}
