--- src/components/Roadmap/roadmap.module.scss
@@ -1,14 +1,0 @@
-.roadmap {
-    &_text {
-        margin-bottom: 24px;
-    }
-
-    &_heading {
-        color: #3A4374;
-        margin-bottom: 4px;
-    }
-
-    &_body {
-        color: #647196;
-    }
-}
