+++ src/layouts/Roadmap/_subcomponents/roadmapmain.module.scss
@@ -1,3 +1,18 @@
 .roadmapMain {
     padding: 24px 24px 56px 24px;
 }
+
+.roadmap {
+    &_text {
+        margin-bottom: 24px;
+    }
+
+    &_heading {
+        color: #3A4374;
+        margin-bottom: 4px;
+    }
+
+    &_body {
+        color: #647196;
+    }
+}
