+++ src/layouts/Roadmap/roadmap.module.scss
@@ -1,0 +1,14 @@
+.roadmap {
+    &_text {
+        margin-bottom: 24px;
+    }
+
+    &_heading {
+        color: #3A4374;
+        margin-bottom: 4px;
+    }
+
+    &_body {
+        color: #647196;
+    }
+}
